Issue #4406 - Moved secure metrics REST handler to its own bundle
