+++ dev/com.ibm.websphere.appserver.features/visibility/public/mpMetrics-1.1/com.ibm.websphere.appserver.mpMetrics-1.1.feature
@@ -13,6 +13,7 @@ Subsystem-Name: MicroProfile Metrics 1.1
  com.ibm.websphere.appserver.servlet-3.1; ibm.tolerates:=4.0,\
  com.ibm.websphere.appserver.mpConfig-1.2; ibm.tolerates:=1.3
 -bundles=com.ibm.ws.require.java8, \
- com.ibm.ws.microprofile.metrics.1.1
+ com.ibm.ws.microprofile.metrics.1.1, \
+ com.ibm.ws.microprofile.metrics.private
 kind=ga
 edition=core
