+++ dev/com.ibm.ws.microprofile.metrics.private/bnd.bnd
@@ -14,53 +14,50 @@ bVersion=1.0
 javac.source: 1.8
 javac.target: 1.8
 
-Bundle-Name: MicroProfile Metrics 1.1
-Bundle-SymbolicName: com.ibm.ws.microprofile.metrics.1.1
+Bundle-Name: MicroProfile Metrics Private REST API
+Bundle-SymbolicName: com.ibm.ws.microprofile.metrics.private
 
-Bundle-Description: MicroProfile Metrics 1.1, version ${bVersion}
+Bundle-Description: MicroProfile Metrics 1.0 Private REST API, version ${bVersion}
 
-IBM-Authorization-Roles: com.ibm.ws.management
 IBM-Web-Extension-Processing-Disabled: true
 
-Import-Package: \
-	javax.enterprise.inject; version="[1.1,3)",\
-	org.eclipse.microprofile.config.*;version="[1.0,2)",\
-	*
+Web-ContextPath: /metrics
 
 Private-Package: \
-	com.ibm.ws.microprofile.metrics.resources,\
-	com.ibm.ws.microprofile.metrics.exceptions,\
-	com.ibm.ws.microprofile.metrics.helper,\
-	com.ibm.ws.microprofile.metrics.writer
+	com.ibm.ws.microprofile.metrics.privateapi,\
+	com.ibm.ws.microprofile.metrics.privateapi.resources.*
 
-Export-Package: \
-	!com.ibm.ws.microprofile.metrics.resources,\
-	com.ibm.ws.microprofile.metrics11.*,\
-	com.ibm.ws.microprofile.metrics.impl,\
-	com.ibm.ws.microprofile.metrics,\
-	com.ibm.ws.microprofile.metrics.helper
-    
+Include-Resource: \
+	WEB-INF=resources/WEB-INF
+
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 
 -dsannotations-inherit: true
--dsannotations: \
-	com.ibm.ws.microprofile.metrics11.impl.SharedMetricRegistries11,\
-	com.ibm.ws.microprofile.metrics.ApplicationListener
+-dsannotations: com.ibm.ws.microprofile.metrics.privateapi.PrivateMetricsRESTHandler
 
 WS-TraceGroup: METRICS
 
 -buildpath: \
+	com.ibm.ws.microprofile.metrics;version=latest,\
 	com.ibm.websphere.org.osgi.core;version=latest,\
 	com.ibm.ws.logging;version=latest,\
 	com.ibm.websphere.org.osgi.service.component;version=latest,\
 	com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
-	com.ibm.websphere.org.eclipse.microprofile.metrics.1.1.1;version=latest,\
-	com.ibm.websphere.org.eclipse.microprofile.config.1.2.1;version=latest,\
+	com.ibm.websphere.org.eclipse.microprofile.metrics.1.0;version=latest,\
+	com.ibm.websphere.appserver.spi.kernel.service;version=latest,\
+	com.ibm.websphere.appserver.api.json;version=latest,\
+	com.ibm.websphere.appserver.spi.httptransport;version=latest,\
+	com.ibm.websphere.javaee.servlet.3.1;version=latest,\
+	com.ibm.websphere.appserver.spi.ssl;version=latest,\
 	com.ibm.wsspi.org.osgi.service.event;version=latest,\
+	com.ibm.websphere.rest.handler;version=latest,\
+	com.ibm.ws.rest.handler;version=latest,\
+	com.ibm.ws.webcontainer;version=latest,\
+	com.ibm.ws.container.service;version=latest,\
+	com.ibm.ws.container.service.compat;version=latest,\
 	com.ibm.websphere.javaee.annotation.1.2;version=latest,\
 	com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
 	com.ibm.websphere.javaee.interceptor.1.2;version=latest,\
-	com.ibm.websphere.javaee.cdi.1.2;version=latest,\
-	com.ibm.ws.microprofile.metrics;version=latest
-
+	com.ibm.websphere.javaee.cdi.1.2;version=latest, \
+	com.ibm.ws.kernel.service;version=latest
