+++ dev/com.ibm.ws.microprofile.metrics.private/resources/WEB-INF/web.xml
@@ -10,21 +10,21 @@
         IBM Corporation - initial API and implementation
  -->
 <!DOCTYPE web-app PUBLIC "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN" "http://java.sun.com/dtd/web-app_2_3.dtd">
-<web-app id="MicroProfileMetricsREST" version="3.0" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd">
-  <display-name>MicroProfileMetrics</display-name>
+<web-app id="PrivateMicroProfileMetricsREST" version="3.0" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd">
+  <display-name>PrivateMicroProfileMetrics</display-name>
   <description>IBM MicroProfile Metrics for Liberty server</description>
 
   <servlet>
-    <servlet-name>MetricsRESTProxyServlet</servlet-name>
-    <servlet-class>com.ibm.ws.microprofile.metrics.MetricsRESTProxyServlet</servlet-class>
+    <servlet-name>PrivateMetricsRESTProxyServlet</servlet-name>
+    <servlet-class>com.ibm.ws.microprofile.metrics.privateapi.PrivateMetricsRESTProxyServlet</servlet-class>
     <load-on-startup>1</load-on-startup>
   </servlet>
   
   <servlet-mapping id="ServletMapping_Default">
-    <servlet-name>MetricsRESTProxyServlet</servlet-name>
+    <servlet-name>PrivateMetricsRESTProxyServlet</servlet-name>
     <url-pattern>/*</url-pattern>
   </servlet-mapping>
-  
+
   <security-constraint id="SecurityConstraint_1">
     <display-name>MicroProfile Metrics Security Constraint</display-name>
     <web-resource-collection id="WebResourceCollection_1">
@@ -49,4 +49,5 @@
     <realm-name>MicroProfileMetrics</realm-name>
   </login-config> 
   
+
 </web-app>
