+++ dev/com.ibm.ws.microprofile.metrics.private/src/com/ibm/ws/microprofile/metrics/privateapi/PrivateMetricsRESTProxyServlet.java
@@ -1,0 +1,17 @@
+/*******************************************************************************
+ * Copyright (c) 2018 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.microprofile.metrics.privateapi;
+
+import com.ibm.ws.microprofile.metrics.BaseMetricsRESTProxyServlet;
+
+public class PrivateMetricsRESTProxyServlet extends BaseMetricsRESTProxyServlet {
+    private static final long serialVersionUID = 1L;
+}
