+++ dev/com.ibm.ws.microprofile.metrics/bnd.bnd
@@ -22,8 +22,6 @@ Bundle-Description: MicroProfile Metrics 1.0, version ${bVersion}
 IBM-Authorization-Roles: com.ibm.ws.management
 IBM-Web-Extension-Processing-Disabled: true
 
-Web-ContextPath: /metrics
-
 Import-Package: \
 	javax.enterprise.inject; version="[1.1,3)",\
 	*
@@ -33,17 +31,16 @@ Private-Package: \
 	com.ibm.ws.microprofile.metrics.exceptions,\
 	com.ibm.ws.microprofile.metrics.helper,\
 	com.ibm.ws.microprofile.metrics.writer
+
 Export-Package: \
 	com.ibm.ws.microprofile.metrics,\
-	com.ibm.ws.microprofile.metrics.impl
-
-Include-Resource: \
-    WEB-INF=resources/WEB-INF
+	com.ibm.ws.microprofile.metrics.impl,\
+	com.ibm.ws.microprofile.metrics.helper
     
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 
 -dsannotations-inherit: true
--dsannotations: com.ibm.ws.microprofile.metrics.MetricsHandler,\
+-dsannotations: \
 	com.ibm.ws.microprofile.metrics.ApplicationListener,\
 	com.ibm.ws.microprofile.metrics.impl.SharedMetricRegistries
